python_engine/requirements.txt
fastapi
uvicorn
llama-cpp-python
pydantic
chromadb
sentence-transformers
torch
easyocr
pymupdf
